+++ examples/vite-react/src/main.tsx
@@ -1,8 +1,8 @@
-import "@miku/annotator-react/register";
+import "@mikuexe/annotator-react/register";
 
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { SourceAnnotator } from "@miku/annotator-react";
+import { SourceAnnotator } from "@mikuexe/annotator-react";
 import { App } from "./App";
 import "./styles.css";